Example of how the codeblock can be used
This is just an example of how the codeblock can be used to highlight features of a certain code, and I think it would be great if we could do something like this throughout. It will ease the transition into color-coded code blocks on the website in a later stage too.